+++ src/pcm.h
@@ -4,8 +4,8 @@
 #include "typedefs.h"
 
 namespace Pcm {
-  i32 decodeOgg(const u8* oggData, u32 oggDataSize, u8** pcmData, u32& pcmDataSize);
-  void resample(u8** pcmData, u32& pcmDataSize, i32 inSampleRate, i32 outSampleRate);
+  i32 decodeOgg(const u8* oggData, u32 oggDataSize, u8** pcmData, u64& pcmDataSize);
+  void resample(u8** pcmData, u64& pcmDataSize, i32 inSampleRate, i32 outSampleRate);
 }
 
 #endif // PCM_H
